Assignment4

+++ webapp/src/main/java/com/csye6225/spring2019/facade/impl/NoteFacadeServiceImpl.java
@@ -11,7 +11,12 @@ import org.apache.logging.log4j.util.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.sql.Timestamp;
 import java.util.List;
 
 @Service
@@ -39,7 +44,7 @@ public class NoteFacadeServiceImpl implements NoteFacadeService {
             if(!isRunLocal() && isAWSURL(x.getUrl())){
                 String keyName = x.getUrl().substring(x.getUrl().indexOf("/"));
                 S3Util.deleteFile(env.getProperty("csye6225.aws.bucket.name"),keyName);
-            }else if(!isAWSURL(x.getUrl())){
+            }else if(!isAWSURL(x.getUrl()) && isRunLocal()){
                 FileUtil.deleteFileFromLocal(x.getUrl());
             }else {
                 log.info("Cannot delete file due to the env"+x.getUrl() );
@@ -57,5 +62,58 @@ public class NoteFacadeServiceImpl implements NoteFacadeService {
         if(Strings.isEmpty(url)) return false;
         String awsSuffix = env.getProperty("csye6225.aws.url.suffix");
         return url.contains(awsSuffix);
+    }
+
+    @Override
+    public void deleteAttachment(String attachmentId){
+        if(Strings.isEmpty(attachmentId)) return;
+        Attachment attachment  = attachmentService.getAttachmentById(attachmentId);
+        if(attachment == null || Strings.isEmpty(attachment.getUrl())) return;
+        String url = attachment.getUrl();
+        deleteFileByUrl(url);
+        attachmentService.deleteAttachmentById(attachmentId);
+    }
+
+
+    private void deleteFileByUrl(String url){
+
+        String bucket = env.getProperty("csye6225.aws.bucket.name");
+        if(isAWSURL(url)){
+            String path = url.substring(url.indexOf(bucket)+bucket.length());
+            S3Util.deleteFile(bucket,path);
+        }else {
+            FileUtil.deleteFileFromLocal(url);
+        }
+    }
+
+    @Override
+    public void updateAttachment(InputStream input, String fileName,String fileType,double fileSize,String attId){
+        if(input==null || Strings.isEmpty(fileName)||Strings.isEmpty(fileType)||Strings.isEmpty(attId)) return;
+        Attachment attachment = attachmentService.getAttachmentById(attId);
+        if(attachment == null ) {
+            log.warn("Cannot find attachment by id :" +attId);
+        }
+        String url;
+        if(isRunLocal()){
+            try {
+                url = FileUtil.saveFileToLocal(input, env.getProperty("csye6225.file.folder"), fileName, fileType);
+            }catch (IOException e){
+                log.error(e);
+                return ;
+            }
+        }else {
+            url = S3Util.uploadFile(env.getProperty("csye6225.aws.bucket.name"),env.getProperty("csye6225.file.folder")
+                    ,input,env.getProperty("csye6225.aws.url.suffix"),fileName,fileType);
+        }
+        if(url ==null){
+            return ;
+        }
+        deleteFileByUrl(attachment.getUrl());
+        attachment.setFileType(fileType);
+        attachment.setUpdateTime(new Timestamp(System.currentTimeMillis()));
+        attachment.setFileSize(fileSize);
+        attachment.setUrl(url);
+        attachment.setFileName(fileName);
+        attachmentService.updateAttachment(attachment);
     }
 }
